% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageProcessing.R
\name{blurImage}
\alias{blurImage}
\title{Apply a blurring filter}
\usage{
blurImage(img, brush_size, brush_shape, sigma = 0.3)
}
\arguments{
\item{img}{An object of Image class specific to EBImage, stored as multi-
dimensional arrays containing the pixel intensities.}

\item{brush_size}{A \code{numeric} value containing the size of the brush
in pixels. This should be an odd number; even numbers are rounded to the
next odd one, i.e., size = 4 has the same effect as size = 5. Default is 5.}

\item{brush_shape}{A \cold{character} vector indicating the shape of the brush.
Can be box, disc, diamond,Gaussian or line. Default is box.}

\item{sigma}{An optional numeric containing the standard deviation of the
Gaussian shape.This argument is relevant only for the
Gaussian brush shape.Default is 0.3.}
}
\description{
The following function uses \code{\link[EBImage]{makeBrush}} to generate
a filter of choice and applies it to the image. Blurring tends to be
useful prior to analyzing individual objects in an image with ill-defined
edges and/or uneven intensities.
}
\examples{
# Example 1
# select the nuclei channel in a tiff file
mouse = readImage(system.file('extdata/Mouse_01.tiff', package='MyoManager'))
nuc = getFrame(mouse,3)

# apply the blurring brush
img_blur = blurImage(nuc, size = 11, shape = 'gaussian', sigma = 5)

# visualize blurred nuclei - brightening is optional (for display purpose)
display(img_blur * 2, method = "raster")

# Example 2
rabbit = readImage(system.file('extdata/Rabbit_01.tiff', package='MyoManager'))


}
\references{
Gregoire Pau, Florian Fuchs, Oleg Sklyar, Michael Boutros, and Wolfgang Huber
(2010): EBImage - an R package for image processing with applications to
cellular phenotypes. \emph{Bioinformatics}, 26(7), pp. 979-981,
\href{https://pubmed.ncbi.nlm.nih.gov/20338898/}{link}
\url{(https://bioconductor.org/packages/release/bioc/html/EBImage.html}
}
