% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleiTool.R
\name{countNuclei}
\alias{countNuclei}
\title{(For multi-frame images), frame selection.}
\usage{
countNuclei(img)
}
\arguments{
\item{img}{An object of Image class specific to EBImage, stored as multi-
dimensional arrays containing the pixel intensities.}
}
\value{
Returns a \code{numeric} value of the number of nuclei counted.
}
\description{
The following function is a wrap around \code{\link[EBImage]{getFrame}} to
select 1 frame to proceed with analysis.
}
\details{
When it comes to counting cellular objects (most commonly nuclei),
The key steps are:
blur the image
apply a threshold to turn nuclei into 'blobs'
count the 'blobs'
}
\examples{
# Example 1
rabbit = readImage(system.file('extdata/Rabbit_01.tif', package='MyoManager'))
rNuc = getFrame(rabbit, 3)
countNuclei(rNuc)
viewImage(rNuc)

# Example 2
mouse = readImage(system.file('extdata/Mouse_01.tif', package='MyoManager'))
mNuc = getFram(mouse, 3)
countNuclei(mNuc)
viewImage(mNuce)

}
\references{
Gregoire Pau, Florian Fuchs, Oleg Sklyar, Michael Boutros, and Wolfgang Huber
(2010): EBImage - an R package for image processing with applications to
cellular phenotypes. \emph{Bioinformatics}, 26(7), pp. 979-981,
\href{https://pubmed.ncbi.nlm.nih.gov/20338898/}{link}
\url{(https://bioconductor.org/packages/release/bioc/html/EBImage.html}
}
