% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleiTool.R
\name{getFeatureData}
\alias{getFeatureData}
\title{Generate a data frame of nuclei shape and size features}
\usage{
getFeatureData(img)
}
\arguments{
\item{img}{A single-frame grayscale \code{Image} containing ONLY the nuclei.}
}
\value{
Returns a \code{data frame} containing area, perimeter, mean radius, and
eccentricity of each nuclei objects in the \code{Image}.
}
\description{
The following function uses \code{computeFunction} and \code{computeFunction.shape}
from \code{EBImage} tp produce a data frame containing the following features:
area, perimeter, mean radius, and eccentricity. ( elliptical eccentricity defined by
sqrt(1-minoraxis^2/majoraxis^2). This value approaches 0 for rounder objects and 1 for
elongated objects.
}
\details{
*Note Frame of only nuclei (typically frame #3 of composite image) is needed, since
feature data extraction can only be performed on single-frame Grayscale images with
defined binary objects.
}
\examples{
# Example 1
rabbit = loadImage(system.file('extdata/Rabbit_01.tif', package='MyoManager'))
rNuc = selectFrame(rabbit, 3)
rNucFeatures_df = getFeatureData(rNuc)

# Example 2
mouse = loadImage(system.file('extdata/Mouse_01.tiff', package='MyoManager'))
mNuc = selectFrame(mouse, 3)
mNucFeatures_df = getFeatureData(mNuc)

}
\references{
Gregoire Pau, Florian Fuchs, Oleg Sklyar, Michael Boutros, and Wolfgang Huber
(2010): EBImage - an R package for image processing with applications to
cellular phenotypes. \emph{Bioinformatics}, 26(7), pp. 979-981,
\href{https://pubmed.ncbi.nlm.nih.gov/20338898/}{link}
\url{https://bioconductor.org/packages/release/bioc/html/EBImage.html}
}
