% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageProcessing.R
\name{intensityCtrl}
\alias{intensityCtrl}
\title{Simple image manipulations}
\usage{
intensityCtrl(img, brightness = 0, contrast = 1)
}
\arguments{
\item{img}{An object of Image class specific to EBImage, stored as multi-
dimensional arrays containing the pixel intensities.}

\item{brightness}{A \code{numeric} value containing the level of brightness to be
applied to the image. Default is 0. (no increased/decreased brightness)}

\item{contrast}{A \code{numeric} value (>0) containing the degree of contrast to be
applied to the image. Default is 1.(no increased/decreased contrast)}
}
\description{
The following function manipulates the brightness and the degree of contrast
of an image.
}
\details{
As EBImage objects are stored as matrices, they can be manipulated with
all R mathematical operators. This includes + to control the brightness,
\itemize{
\item to control the degree of contrast, or ^ to control the gamma correction parameter.
}
}
\examples{
# Example 1: decrease brightness and increase contrast
# (good for object identification)
mouse = readImage(system.file('extdata/Mouse_01.tiff', package='MyoManager'))
mNuc = selectFrame(mouse, 3)
viewImage(mNuc)

mouse_enhanced = intensityCtrl(mNuc, -0.2, 3)
viewImage(mouse_enhanced)

# Example 2: increase brightness and decrease contrast
human =readImage(system.file('extdata/Human_01.tiff', package='MyoManager'))
hNuc = selectFrame(human, 3)
viewImage(hNuc)

human_enhanced = intensityCtrl(hNuc, -0.2, 3)
viewImage(human_enhanced)

}
