% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageProcessing.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Private Helper}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A numeric value to be tested upon.}
}
\description{
R's base function is.integer(x) does not test if x contains integer numbers.
Hence the added helper to test for integer.
}
\references{
R Core Team (2021). R: A language and environment for statistical computing. R
Foundation for Statistical Computing, Vienna, Austria.
\href{https://www.R-project.org/}{link}.
}
