% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageProcessing.R
\name{makePretty}
\alias{makePretty}
\title{Simple image manipulations}
\usage{
makePretty(img, brightness, constrast)
}
\description{
The following function manipulates the brightness and the degree of contrast
of an image.
}
\details{
As EBImage objects are stored as matrices, they can be manipulated with
all R mathematical operators. This includes + to control the brightness,
\itemize{
\item to control the degree of contrast, or ^ to control the gamma correction parameter.
}
}
\examples{
# Example 1

# Example 2

}
\references{

}
