% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleiTool.R
\name{plotFeature}
\alias{plotFeature}
\title{Plot a shape/size feature of nuclei objects.}
\usage{
plotFeature(featureDF, feature = c("area", "perimeter", "radius", "roundness"))
}
\arguments{
\item{featureDF}{A \code{data frame} containing four columns of shape/size features for
each nuclei objects in an \code{Image}.}

\item{feature}{A \code{character} vector indicating which feature to plot, can be
area, perimeter, radius, or roundness.}
}
\description{
The following function plots the density distribution of one of four shape/size features
of cell nuceli computed from getFeatureData(): area, perimeter, radius, or eccentricity.
}
\examples{
\dontrun{
# Example 1
rabbit = loadImage(system.file('extdata/Rabbit_01.tif', package='MyoManager'))
rNuc = selectFrame(rabbit, 3)
df = getFeatureData(rNuc)
plotFeature(df, "area")

# Example 2
plotFeature(df, "roundness")
}
}
\references{
Gregoire Pau, Florian Fuchs, Oleg Sklyar, Michael Boutros, and Wolfgang Huber
(2010): EBImage - an R package for image processing with applications to
cellular phenotypes. \emph{Bioinformatics}, 26(7), pp. 979-981,
\href{https://pubmed.ncbi.nlm.nih.gov/20338898/}{link}
\url{https://bioconductor.org/packages/release/bioc/html/EBImage.html}
}
