% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMyoManager.R
\name{runMyoManager}
\alias{runMyoManager}
\title{Launch Shiny App For Package MyoManager}
\usage{
runMyoManager()
}
\value{
No return value but open up a shiny page.
}
\description{
A function that launches the shiny app for this package.
The shiny app allows users to perform microscopy image
analysis of muscle stem cells and other similar cell types.
The actual script has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
runMyoManager()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
\author{
Yinni Kuang, \email{kareen.kuang@mail.utoronto.ca}
}
