% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentImage.R
\name{segmentImage}
\alias{segmentImage}
\title{Generate segmented images of cellular structures}
\usage{
segmentImage(
  img,
  cell_frame,
  nuc_frame,
  show_structure = c("cell", "nuclei", "both")
)
}
\arguments{
\item{img}{An object of Image class specific to EBImage, stored as multi-
dimensional arrays containing the pixel intensities.}

\item{cell_frame}{A \code{numeric} value indicating the frame in the composite image
that contains stained cell (fiber)}

\item{nuc_frame}{A \code{numeric} value indicating the frame in the composite image
that contains stained nuclei}

\item{show_structure}{A \code{character} vector specifying which segmented structure
to show. Can be cell, nuclei, or both (segmented nuclei overlaying segmented cells).
Default is both.}
}
\value{
Returns an \code{Image} containing the segmented structure(s).
}
\description{
The following function uses several functions from \code{EBImage} to
perform image segmentation with the provided frame numbers of cell and nuclei
channels.
}
\details{
Segmentation of individual image channels is often useful when it is difficult
to distinguish structures by eye. (e.g. Visualizing segmentation helps users
locate nuclei within a mesh of muscle fibers)
}
\examples{
# Example 1
rabbit <- loadImage(system.file('extdata/Rabbit_01.tif', package = 'MyoManager'))
# view image in Grayscale to find out which Frame stores cell body / nuclei
viewImage(rabbit, Grayscale)
segmented_nuc <- segmentImage(rabbit, 1, 3, 'nuclei')
# only nuclei segmentation is generated
viewImage(segmented_nuc)


# Example 2
mouse <- loadImage(system.file('extdata/Mouse_01.tiff', package = 'MyoManager'))
viewImage(mouse, Grayscale)
segmented_nuc <- segmentImage(mouse, 2, 3)
# cell segmentation with highlighted nuclei is generated
viewImage(segmented_nuc)

}
\references{
Gregoire Pau, Florian Fuchs, Oleg Sklyar, Michael Boutros, and Wolfgang Huber
(2010): EBImage - an R package for image processing with applications to
cellular phenotypes. \emph{Bioinformatics}, 26(7), pp. 979-981,
\href{https://pubmed.ncbi.nlm.nih.gov/20338898/}{link}
\url{https://bioconductor.org/packages/release/bioc/html/EBImage.html}

Carpenter, A.E., Jones, T.R., Lamprecht, M.R. et al. (2006) CellProfiler: image
analysis software for identifying and quantifying cell phenotypes. \emph{Genome Biol}
R100:7. \href{https://doi.org/10.1186/gb-2006-7-10-r100}{link}
CellProfiler software: \url{http://www.cellprofiler.org}
}
