% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageProcessing.R
\name{selectFrame}
\alias{selectFrame}
\title{Frame selection.}
\usage{
selectFrame(img, frame_number)
}
\arguments{
\item{img}{An object of Image class specific to EBImage, stored as multi-
dimensional arrays containing the pixel intensities.}

\item{frame_number}{A \code{numeric} value specifying a frame in img.}
}
\value{
A single-frame grayscale \code{Image} indicated by frame_number.
}
\description{
The following function selects 1 frame from a composite image to proceed with analysis.
}
\details{
Microscopy data typically comes in multiple frames (or channels), each showing
a different color of fluorescence/immuno-staining (e.g. nuclei, cell body, protein
markers). The composite image is often shown in colored to differentiate cellular
structures, while individual frames are typically shown in greyscale.
}
\examples{
# Example 1
# visualize the nuclei channel in a tiff file
human = loadImage(system.file('extdata/Human_01.tiff', package='MyoManager'))
nuc = selectFrame(human, 3)
viewImage(nuc)


# Example 2
# visualize the fiber(cell body) channel in a jpg image
mouse = loadImage(system.file('extdata/Mouse_02.tif', package='MyoManager'))
fib = selectFrame(mouse,2)
viewImage(nuc)

}
\references{
Gregoire Pau, Florian Fuchs, Oleg Sklyar, Michael Boutros, and Wolfgang Huber
(2010): EBImage - an R package for image processing with applications to
cellular phenotypes. \emph{Bioinformatics}, 26(7), pp. 979-981,
\href{https://pubmed.ncbi.nlm.nih.gov/20338898/}{link}
\url{(https://bioconductor.org/packages/release/bioc/html/EBImage.html}
}
