% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndDisplay.R
\name{ValidImage}
\alias{ValidImage}
\title{Private Helper}
\usage{
ValidImage(image_obj)
}
\arguments{
\item{image_obj}{An object of Image class specific to EBImage, stored as multi-
dimensional arrays containing the pixel intensities.}
}
\description{
checks whether 'i' is a suitable image
}
