% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndDisplay.R
\name{LoadImage}
\alias{LoadImage}
\title{Loading and displaying images}
\usage{
LoadImage(image_path)
}
\arguments{
\item{image_path}{A \code{character} vector of one or more paths to image files}
}
\value{
LoadImage does not return. ViewImage displays an image object
processed by LoadImage using R graphics.
}
\description{
The following functions handles several types of image data (jpeg, png, tiff)
by importing images and prompting a R graphics window for display.
The graphics device supports multiple frames and different degrees of zoom,
which makes it easy to visualize detailed microscopy data.
}
\examples{
# Example 1

\dontrun{
# Example 2

}
}
\references{
Jeroen Ooms (2021). magick: Advanced Graphics and Image-Processing in R. R
package version 2.7.3. \href{https://CRAN.R-project.org/package=magick}{link}

Gregoire Pau, Florian Fuchs, Oleg Sklyar, Michael Boutros, and Wolfgang Huber
(2010): EBImage - an R package for image processing with applications to
cellular phenotypes. \emph{Bioinformatics}, 26(7), pp. 979-981,
\href{https://pubmed.ncbi.nlm.nih.gov/20338898/}{link}
\url{(https://bioconductor.org/packages/release/bioc/html/EBImage.html}
}
