% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndDisplay.R
\name{loadImage}
\alias{loadImage}
\title{Loading images}
\usage{
loadImage(image_path)
}
\arguments{
\item{image_path}{A \code{character} vector of one or more paths or URLs to image files}
}
\value{
An object of \code{Image} class specific to EBImage, stored as multi-
dimensional arrays containing the pixel intensities.
}
\description{
The following function is a wrap around \code{\link[magick]{image_read}} and
\code{\link[magick]{as_EBImage}} to load one or more images from files.
}
\details{
Magick can handle for a variety of image formats (over 200), EBImage has a
superior display quality but only supports the three common image formats
jpg, png, and tiff. Hence by loading images with Magick and converting them
to EBImage objects for display, we can avoid unsupported image types while
maximizing display functionalities.
}
\examples{
# Example 1
#load sample tiff distributed with the package
library(MyoManager)
image <- loadImage(system.file('extdata/Human_01.tiff', package = 'MyoManager'))


# Example 2
#load sample jpeg from public link
library(MyoManager)
image <- loadImage("https://user-images.githubusercontent.com/60583839/141215629-f19d4a77-c5f0-491f-9262-b22cd59739e3.jpg")


}
\references{
Jeroen Ooms (2021). magick: Advanced Graphics and Image-Processing in R. R
package version 2.7.3. \href{https://CRAN.R-project.org/package=magick}{link}

Gregoire Pau, Florian Fuchs, Oleg Sklyar, Michael Boutros, and Wolfgang Huber
(2010): EBImage - an R package for image processing with applications to
cellular phenotypes. \emph{Bioinformatics}, 26(7), pp. 979-981,
\href{https://pubmed.ncbi.nlm.nih.gov/20338898/}{link}
\url{(https://bioconductor.org/packages/release/bioc/html/EBImage.html}
}
