% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndDisplay.R
\name{ViewImage}
\alias{ViewImage}
\title{Displays image}
\usage{
ViewImage(image_obj)
}
\arguments{
\item{image_obj}{An object of Image class specific to EBImage, stored as multi-
dimensional arrays containing the pixel intensities.}
}
\description{
Prompts R's display window to open for single-frame and multi-frame images
}
